No need for multi-release (11, 16) when project is 17
Collapsed MR jar and compilation, by keeping latest
classes.
